% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_footfall.R
\name{fit_footfall}
\alias{fit_footfall}
\title{Linear regression model for Dublin 2021 footfall data.}
\usage{
fit_footfall(dat, data_type = "daily")
}
\arguments{
\item{dat}{The Dublin 2021 footfall data, from the \code{"load_footfall()"}
function.}

\item{data_type}{Whether the model should examine daily, weekly
or monthly data. Set to daily by default.}
}
\value{
An object of class \code{"lrmodel"} which contains the attributes:
mod.residuals: residual values,
mod.fitted: fitted values,
mod.coef: coefficients,
mod.summary: model summary,
mod.fstat: F-statistic,
mod.call: the call, ie. the formula used to generate the model,
mod.sigma: the sigma value.
}
\description{
The function creates a linear regression model for either daily, weekly or
monthly data for the Dublin 2021 footfall data set. It returns an object
containing a variety of values of potential interest.
}
\examples{
dublin <- load_footfall()
fit1 <- fit_footfall(dublin)
fit2 <- fit_footfall(dublin, data_type = "weekly")
fit3 <- fit_footfall(dublin, data_type = "monthly")

plot(fit1$mod.fitted, fit1$mod.residuals)
fit2$mod.sigma
fit3$mod.summary
}
\seealso{
\code{\link{load_footfall}}, \code{\link{plot_footfall}}
}
\author{
Emer Phelan - </{emer.phelan.2019@mumail.ie}>
}
