% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_footfall.R
\name{load_footfall}
\alias{load_footfall}
\title{Load in function for footfall data for the year 2021.}
\usage{
load_footfall(date_format = "\%d-\%m-\%Y", time_format = "\%R", na.rm = TRUE)
}
\arguments{
\item{date_format}{The default date_format is now in
day-month-year, this can be changed to have month as
January. This can be changed instead of it showing this
as numeric 01, one can replace the lowercase letter m to
the lowercase letter b.This can be seen below
 01 which can be}

\item{time_format}{The default for the time_format is in
specifying R which is the equivalent to hour - minutes. While
replacing this with T in the code below is adding seconds which
is equivalent to hours, minutes and seconds.}

\item{na.rm}{This argument is equal to TRUE whereby it removes
any NA values within the data itself.}
}
\value{
An object of class footfall that has four columns
Date, Time, Location and Count.

The date is in format day-month-year
The time column is in hours and minutes
The locations columns gives the locations
of various streets around the city.
The count column sums up the amount of
people in the area at dates and time as listed.
}
\description{
This functions imports 2021 footfall data from the smart
Dublin website and cleans this up too be used later on.
}
\examples{
date_format1 <- load_footfall("\%d-\%m-\%Y)

date_format2 <- load_footfall("\%d-\%b-\%Y)

time_format1 <-load_footfall(time_format = "\%R")

time_format2 <- load_footfall(time_format = "\%T")



}
\seealso{
See \code{\link{strptime}} for the available conversion
specifications for the \code{date_format} and \code{time_format}
}
\author{
Liam Padden - <\email{liam.padden.2020@mumail.ie}
}
