% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_footfall.R
\name{plot_footfall}
\alias{plot_footfall}
\title{Visualisation for Dublin 2021 footfall data.}
\usage{
plot_footfall(dat, mod, plot_type = "res vs fit")
}
\arguments{
\item{dat}{The Dublin 2021 footfall data, from the \code{"load_footfall()"}
function.}

\item{mod}{An object of type \code{"lrmodel"}, from the
\code{"fit_footfall()"} function. If no object is provided, one is created
by default.}

\item{plot_type}{The type of plot produced. Can be "mean" for the mean
footfall plot, "total" for the total footfall plot or "res vs fit" for the
residual vs fitted plot. Set to residual vs fitted plot by default.}
}
\value{
One of three plots:
A scatter plot of mean footfall counts across each Dublin location,
A scatter plot of total footfall counts across each Dublin location,
A residual vs fitted plot from the linear model object.
}
\description{
The function creates multiple plots relating to the Dublin 2021 footfall
data set. It creates one of three plots: a scatter plot with mean footfall
counts, a scatter plot with total footfall counts and a residual vs fitted
plot for a linear regression created from the data.
}
\examples{
dublin <- load_footfall()
fit1 <- fit_footfall(dublin)

plot_footfall(dublin, fit1)
plot(dublin, fit1, plot_type = "mean")
}
\seealso{
\code{\link{load_footfall}}, \code{\link{fit_footfall}}
}
\author{
Obinna Njoku - </{Obinna.Njoku.2019@mumail.ie}>
}
